requirements.txt
appdirs==1.4.3
asgiref==3.3.4
beautifulsoup4==4.9.3
CacheControl==0.12.6
certifi==2019.11.28
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.2.4
django-bootstrap4==3.0.1
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-filter==2.4.0
django-heroku==0.3.1
django-js-asset==1.2.2
djangorestframework==3.12.4
gunicorn==20.1.0
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
Markdown==3.3.4
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
Pillow==8.2.0
progress==1.5
psycopg2==2.8.6
pyparsing==2.4.6
pytoml==0.1.21
pytz==2021.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
soupsieve==2.2.1
sqlparse==0.4.1
toml==0.10.2
urllib3==1.25.8
webencodings==0.5.1
whitenoise==5.2.0
